fix: separate data structures for questions and GORC
