+++ src/data-types/GROCNodes.ts
@@ -1,25 +1,27 @@
+export type GORCNode = EssentialElement | Category | Subcategory | Attribute | Feature;
+
 export type EssentialElement = IdentifiableEntity & Attributes & {
     type: "essential-element",
     categories: Category[];
-} & Questions;
+};
 
 export type Category = IdentifiableEntity & Attributes & {
     type: "category",
     subcategories: Subcategory;
-} & Questions;
+};
 
 export type Subcategory = IdentifiableEntity & Attributes & {
     type: "subcategory",
-} & Questions;
+};
 
 export type Attribute = IdentifiableEntity & KPIs & {
     type: "attribute",
     features: Feature[];
-} & Questions;
+};
 
 export type Feature = IdentifiableEntity & KPIs & {
     type: "feature",
-} & Questions;
+};
 
 export type KPI = IdentifiableEntity & {
     type: "kpi",
@@ -42,25 +44,15 @@ type Attributes = {
     attributes: Attribute[];
 }
 
-type Questions = {
-    questions: Question[];
-}
-
 type KPIs = {
     kpis: KPI[];
 }
 
-export type Source = {
+type Source = {
     name: string;
     url: string;
 }
 
-export type ConsiderationLevel = "core" | "desirable" | "optional";
+type ConsiderationLevel = "core" | "desirable" | "optional";
 
-export type Question = {
-    id: string;
-    label: string;
-    description: string;
-}
-
-export type Implementation = unknown;
+type Implementation = unknown;
